New: Support for column templates, providing the ability to use Vue components in DataTables' columns. This new ability requires DataTables 2 (still in development at the time of this commit)
Use a example with the slot name `column-X` where `X` is the column
index that you want the template to apply to. DataTables will render the
template via Vue's methods, and insert it into the table for each cell
in that column. This is implemented by using a rendering function for
the column. DataTables 2's ability to have an HTML element returned
from the rendering function is key here.